QR_Code_Processor/requirements.txt
alembic==1.7.5
aniso8601==6.0.0
appdirs==1.4.3
apturl==0.5.2
astroid==2.2.5
attrs==19.3.0
awscli==1.22.47
Babel==2.9.1
backports.entry-points-selectable==1.1.1
bcrypt==3.1.7
blinker==1.4
botocore==1.23.47
Brlapi==0.7.0
cached-property==1.5.1
ccy==1.2.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.12.0
ecdsa==0.13.2
email-validator==1.1.3
entrypoints==0.3
fasteners==0.14.1
filelock==3.3.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.5.0
Flask-WTF==1.0.0
future==0.17.1
greenlet==1.1.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
migrate==0.3.8
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pbr==5.8.0
pexpect==4.6.0
Pillow==9.0.0
pipenv==11.9.0
platformdirs==2.4.0
pluggy==0.13.1
postgres==4.0
protobuf==3.6.1
psycopg2==2.9.1
psycopg2-binary==2.8.2
psycopg2-pool==1.1
py==1.11.0
pyasn1==0.4.8
pycairo==1.16.2
pycryptodome==3.3.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==6.2.2
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-jose==3.3.0
pytz==2019.1
pyxdg==0.26
PyYAML==5.3.1
pyzbar==0.1.8
qrcode==7.3.1
redis==4.0.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.7.2
s3transfer==0.5.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.4
stevedore==3.5.0
systemd-python==234
texttable==1.6.2
toml==0.10.2
typed-ast==1.4.2
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==0.15.6
wrapt==1.11.1
WTForms==3.0.0
xkit==0.0.0
zipp==3.6.0
